fix compilation under 32bit

+++ vod/subtitle/cap_format.c
@@ -100,7 +100,7 @@ cap_get_duration(vod_str_t* source)
 			continue;
 		}
 
-		if (len > end - p)
+		if (len > (uint32_t)(end - p))
 		{
 			break;
 		}
@@ -167,7 +167,7 @@ cap_get_next_block(u_char* p, u_char* end)
 			continue;
 		}
 
-		if (len > end - p)
+		if (len > (uint32_t)(end - p))
 		{
 			break;
 		}
